% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{aap}
\alias{aap}
\title{All available portfolio}
\usage{
aap(returns, rg = 0.01, rfr = 0.001, precision = 0.01, mirc = 1, ...)
}
\arguments{
\item{returns}{xts or zoo value}

\item{rg}{goal of returns}

\item{rfr}{risk free rate}

\item{precision}{grid precision about investment weight}

\item{mirc}{maximum investment ratio constraint}
}
\description{
Return of all available portfolio
}
\examples{
returns <- xdiff_returns(sample_index, 1)
aap(returns, rg = .01, rfr = .001)
}

