% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess_rank}
\alias{assess_rank}
\title{Ranking assess funds using returns xts dataset}
\usage{
assess_rank(returns.xts, choice.stock, step = F, ...)
}
\arguments{
\item{returns.xts}{an \code{xts} object about returns dataset}

\item{method}{returns mean or sd or Sharpe ratio}
}
\value{
data.frame type about assessment result
}
\description{
Ranking assess funds using returns xts dataset & rank plotting
}
\examples{
sample_index \%>\% xdiff_returns(2) \%>\% assess_rank("Fund_2")
sample_index \%>\% xdiff_returns(2) \%>\% assess_rank("Fund_2", step = T)
}
