% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess_returns}
\alias{assess_returns}
\title{Assess funds using returns xts dataset}
\usage{
assess_returns(returns.xts, method = c("mean", "sd", "S.R"), ...)
}
\arguments{
\item{returns.xts}{an \code{xts} object about returns dataset}

\item{method}{returns mean or sd or Sharpe ratio}
}
\value{
data.frame type about assessment result
}
\description{
Assess funds using returns xts dataset
}
\examples{
returns <- sample_index \%>\% xdiff_returns(1)

assess_returns(returns, "mean")
assess_returns(returns, "sd")
assess_returns(returns, "S.R")
}

