% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat}
\alias{cormat}
\title{Visualization of correlation plot using psych::pairs.panels() function}
\usage{
cormat(dat, pch = ".", ...)
}
\description{
This function is alias \code{psych::pairs.panels()}. It can visualize for standardization.
}
\examples{
cormat(cars)
}

