% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{csv2rda}
\alias{csv2rda}
\title{Write rda about imported tables}
\usage{
csv2rda(path, ...)
}
\arguments{
\item{path}{a character vector; file paths}
}
\description{
외부테이블 파일(.txt, .csv 포맷)들을 import 한 후 rda 형태로 모두 변환하여 저장합니다. 객체명은 \code{rawdata} 로 저장됩니다.
}
