% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr}
\alias{descr}
\title{Describe dataset using psych::describe() function}
\usage{
descr(x, choice = c("n", "mean", "sd", "median", "min", "max", "range"),
  ...)
}
\arguments{
\item{x}{A data frame or matrix}

\item{...}{Arguments to be passed to \code{psych::describe()}}
}
\description{
Alias function of \code{psych::describe()}
}
\examples{
mtcars
descr(mtcars)
descr(mtcars, c("n", "mean", "range"))
}
