% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{diff_returns}
\alias{diff_returns}
\title{Accounting x diff returns}
\usage{
diff_returns(xts, x, log.returns = F, start.omit = T, ...)
}
\arguments{
\item{xts}{an \code{xts} object}

\item{x}{an integer vector}

\item{log.returns}{logical asking whether to calculate log returns}
}
\value{
xts type about returns dataset
}
\description{
Accounting x diff returns
}
\examples{
# rolling returns
diff_returns(sample_index, x = 1)
diff_returns(sample_index, x = 1, log.returns = T)

# changed rolling returns
diff_returns(sample_index, x = 0:19)
diff_returns(sample_index, x = 0:19, log.returns = T)
}

