% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{efff}
\alias{efff}
\title{Efficient Frontier}
\usage{
efff(returns, rg = 0.01, rfr = 0.001, short = "no",
  max.allocation = NULL, risk.premium.up = 0.9,
  risk.increment = 1e-04, plot.only.efff = F, ...)
}
\arguments{
\item{returns}{xts or zoo value}

\item{rg}{numeric value, returns of goal}
}
\description{
Efficient Frontier using excess returns
}
\examples{
returns <- xdiff_returns(sample_index, x = 1)
efff(returns, rg = .01, rfr = .001)
}
