% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librarys.R
\name{librarys}
\alias{librarys}
\title{Loading of many packages (using vector type) before auto update packages}
\usage{
librarys(packages, auto.upd = TRUE, auto.ins = FALSE, message = TRUE)
}
\arguments{
\item{packages}{ë¶ˆëŸ¬<U+FFFD><U+FFFD> <U+FFFD>¨í‚¤ì§€ë¥<U+FFFD> vector <U+FFFD>•ìœ¼ë¡<U+FFFD> <U+FFFD>‘ì„±<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>. ë¬¸ì<U+FFFD>´ë¡œ <U+FFFD>‘ì„± <U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}

\item{auto.upd}{<U+FFFD>¨í‚¤ì§€ë¥<U+FFFD> ë¶ˆëŸ¬<U+FFFD>¤ê¸° <U+FFFD>„ì— \code{update.packages(ask=F)} ë¥<U+FFFD> <U+FFFD>¤í–‰<U+FFFD>œì¼œ ëª¨ë“  <U+FFFD>¨í‚¤ì§€<U+FFFD><U+FFFD> <U+FFFD>…ê·¸<U+FFFD>ˆì´<U+FFFD>œë<U+FFFD> <U+FFFD>˜í–‰<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>. <U+FFFD>ì¹˜ <U+FFFD>ŠëŠ”<U+FFFD>¤ë©´ \code{auto.upd=FALSE} ë¥<U+FFFD> <U+FFFD>…ë ¥<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}

\item{auto.ins}{ë§Œì•½ ë¶ˆëŸ¬<U+FFFD><U+FFFD> <U+FFFD>¨í‚¤ì§€ ì¤‘ì— <U+FFFD>¤ì¹˜ê°€ <U+FFFD>˜ì–´<U+FFFD>ˆì<U+FFFD> <U+FFFD>Šì<U+FFFD> <U+FFFD>¨í‚¤ì§€ê°€ <U+FFFD>ˆë‹¤ë©<U+FFFD> <U+FFFD>ë™<U+FFFD>¼ë¡œ <U+FFFD>¤ì¹˜<U+FFFD><U+FFFD> ê²ƒì¸ì§€ <U+FFFD>¬ë<U+FFFD>ë¥<U+FFFD> <U+FFFD>•í•©<U+FFFD>ˆë‹¤.}

\item{message}{<U+FFFD>¨í‚¤ì§€ ë¡œë“œ ê²°ê³¼ summary ë¥<U+FFFD> ì¶œë ¥<U+FFFD><U+FFFD> ê²ƒì¸ì§€ <U+FFFD>¬ë<U+FFFD>ë¥<U+FFFD> <U+FFFD>•í•©<U+FFFD>ˆë‹¤.}
}
\value{
summary ê²°ê³¼ë¥<U+FFFD> data.frame <U+FFFD>•ìœ¼ë¡<U+FFFD> ë°˜í™˜<U+FFFD>˜ë©°, \code{message=FALSE} <U+FFFD><U+FFFD> ë°˜í™˜<U+FFFD>˜ëŠ” ê²ƒì<U+FFFD> <U+FFFD>†ìŠµ<U+FFFD>ˆë‹¤.
}
\description{
2ê°<U+FFFD> <U+FFFD>´ìƒ<U+FFFD><U+FFFD> Package<U+FFFD>¤ì„ <U+FFFD>œë²ˆ<U+FFFD><U+FFFD> ë¡œë“œ<U+FFFD>˜ë©°, <U+FFFD>™ì‹œ<U+FFFD><U+FFFD> <U+FFFD>…ë°<U+FFFD>´íŠ¸ ê°€<U+FFFD><U+FFFD> <U+FFFD>¨í‚¤ì§€ë¥<U+FFFD> <U+FFFD>…ë°<U+FFFD>´íŠ¸ <U+FFFD><U+FFFD> <U+FFFD><U+FFFD> <U+FFFD>´ì „ë²„ì „ê³<U+FFFD> <U+FFFD>¤ì¹˜<U+FFFD><U+FFFD> ë²„ì „<U+FFFD><U+FFFD> <U+FFFD>”ì•½<U+FFFD>˜ì—¬ ì¶œë ¥<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.
}
\examples{
# <U+653C><U+3E62>‹¤ì¤‘ì˜ <U+653C><U+3E64>Œ¨<U+653C><U+3E64>‚¤ì§€ <U+653C><U+3E62>‹¨<U+653C><U+3E63>ˆœ ë¡œë“œ
librarys(packages=c("abind", "acepack", "animation", "arules", "ash"), auto.upd=FALSE)
# ê¸<U+623C><U+3E30> <U+653C><U+3E63>„¤ì¹˜ëœ ëª¨ë“ <U+653C><U+3E64>Œ¨<U+653C><U+3E64>‚¤ì§€ <U+653C><U+3E63>—…ê·¸ë ˆ<U+653C><U+3E63>´<U+653C><U+3E62>“œ <U+653C><U+3E63>´<U+653C><U+3E64>›„ <U+653C><U+3E62>‹¤ì¤<U+393C><U+3E31> <U+653C><U+3E64>Œ¨<U+653C><U+3E64>‚¤ì§€ ë¡œë“œ
librarys(packages=c("abind", "acepack", "animation", "arules", "ash"))
# ë¯¸ì„¤ì¹<U+393C><U+3E38> <U+653C><U+3E64>Œ¨<U+653C><U+3E64>‚¤ì§€ <U+653C><U+3E63>„¤ì¹<U+393C><U+3E38> ë°<U+383C><U+3E66> <U+653C><U+3E63>—…ê·¸ë ˆ<U+653C><U+3E63>´<U+653C><U+3E62>“œ <U+653C><U+3E63>´<U+653C><U+3E64>›„ <U+653C><U+3E62>‹¤ì¤<U+393C><U+3E31> <U+653C><U+3E64>Œ¨<U+653C><U+3E64>‚¤ì§€ ë¡œë“œ
librarys(packages=c("abind", "acepack", "animation", "arules", "ash"), auto.ins=TRUE)
}
\seealso{
\link{library}
}

