% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_col.R
\name{lt_col}
\alias{lt_col}
\title{Color Palettes}
\usage{
lt_col(n, alpha = 1, random = F, ...)
}
\arguments{
\item{n}{the number of colors (<U+FFFD><U+FFFD> 1) to be in the palette.}

\item{alpha}{the alpha transparency, a number in [0,1].}
}
\description{
Color Palettes for kyobo
}
\examples{
(colors1 <- lt_col(9, alpha = .5))
pie(rep(1, 9), col = colors1, clockwise = T)
}

