% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoint.R
\name{midpoint}
\alias{midpoint}
\title{midpoint values}
\usage{
midpoint(x, ...)
}
\arguments{
\item{x}{가공의 대상이 되는 numeric type 벡터를 입력합니다.}

\item{...}{\code{lead()} 함수의 상속 부분입니다.}
}
\description{
실수 벡터에 대해 midpoint 를 반환합니다.
}
\examples{
midpoint(1:10)
}

