% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missmap.R
\name{missmap2}
\alias{missmap2}
\title{Missingness Map using ggplot2}
\usage{
missmap2(x, only_miss_var = F)
}
\arguments{
\item{x}{A dataframe or matrix}
}
\description{
Plots a Missingness Map showing using ggplot2
}
