% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_paste0.R
\name{NA_paste0}
\alias{NA_paste0}
\title{Ignore NA paste function}
\usage{
NA_paste0(...)
}
\description{
ê²°ì¸¡ì¹<U+FFFD> NA ê°€ ì¡´ì¬<U+FFFD><U+FFFD> <U+FFFD><U+FFFD>, NAë¥<U+FFFD> ë¬¸ì<U+FFFD>´ë¡œ <U+FFFD>¸ì‹<U+FFFD>˜ì<U+FFFD> <U+FFFD>Šê³  ê³µë<U+FFFD><U+FFFD>¼ë¡œ paste ë¥<U+FFFD> <U+FFFD>˜í–‰<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.
}
\examples{
### ex1
NA_paste0(c(3,6,NA,NA), c(NA,6,9,NA))

### ex2
A <- "Hello "
B <- c("world", "world!", NA)
paste0(A, B)
NA_paste0(A, B)
}
\seealso{
\link{paste0}
}

