% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendir.R
\name{opendir}
\alias{opendir}
\title{open the working directory directly on R console}
\usage{
opendir(dir = getwd())
}
\arguments{
\item{dir}{default is \code{getwd()}}
}
\description{
open the working directory directly on R console
}
\examples{
opendir()
}

