% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfd.R
\name{openfd}
\alias{openfd}
\title{Open the file or directory directly on R console}
\usage{
openfd(path = NULL)
}
\arguments{
\item{path}{character, default is \code{getwd()}}
}
\description{
Open the file or directory directly on R console
}
\examples{
openfd()
}
