% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste2.R
\name{paste2}
\alias{paste2}
\title{Ignore NA paste function}
\usage{
paste2(...)
}
\description{
ê²°ì¸¡ì¹<U+FFFD> NA ê°€ ì¡´ì¬<U+FFFD><U+FFFD> <U+FFFD><U+FFFD>, NAë¥<U+FFFD> ë¬¸ì<U+FFFD>´ë¡œ <U+FFFD>¸ì‹<U+FFFD>˜ì<U+FFFD> <U+FFFD>Šê³  ê³µë<U+FFFD><U+FFFD>¼ë¡œ paste ë¥<U+FFFD> <U+FFFD>˜í–‰<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.
}
\examples{
A <- "Hello "
B <- c("world", "world!", NA)
paste0(A, B)
paste2(A, B)
}
\seealso{
\link{paste0}
}

