% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_waffle.R
\name{prop_waffle}
\alias{prop_waffle}
\title{Waffle chart (Add proportions)}
\usage{
prop_waffle(x, rounding = 1, ...)
}
\description{
This function that transforms the \code{waffle::waffle()}. It displays the proportions in the legend text.
}
\examples{
x <- c("Apple" = 1, "Banana" = 2, "Watermelon" = 3)
prop_waffle(x, 1, 4, size = 1)
}
