% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psum.R
\name{psum}
\alias{psum}
\title{Returns the parallel sum of the input values.}
\usage{
psum(..., na.rm = F)
}
\arguments{
\item{...}{input values}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\description{
Returns the parallel sum of the input values.
}
\examples{
psum(1:5, c(1, 2, 3, 4, NA), c(1, NA, 3, 4, 5))
psum(1:5, c(1, 2, 3, 4, NA), c(1, NA, 3, 4, 5), na.rm=T)
}

