% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{range01}
\alias{range01}
\title{Range scaling 0 to 1}
\usage{
range01(x, ...)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Range scaling 0 to 1
}
\examples{
x <- c(0, 4, 7, 9, 10)
range01(x)
range01(x) * 100 # Standardization 0 to 100
}
