% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_week.R
\name{recent_weekday}
\alias{recent_weekday}
\title{Sequence Date Generation (only Weekday) from recent day}
\usage{
recent_weekday(recent_day, to = Sys.Date())
}
\arguments{
\item{recent_day}{ëª‡ì¼ì°¨ë<U+FFFD> <U+FFFD>´ê±°<U+FFFD><U+FFFD> ê²ƒì¸ì§€ <U+FFFD>•í•©<U+FFFD>ˆë‹¤. <U+FFFD>°ì´<U+FFFD><U+FFFD> <U+FFFD>•ì‹<U+FFFD>€ <U+FFFD>•ìˆ˜<U+FFFD>´ì–´<U+FFFD><U+FFFD> <U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}

\item{to}{<U+FFFD>¤ëŠ˜<U+FFFD>¼ìê¸°ì<U+FFFD><U+FFFD><U+FFFD> <U+FFFD>•í•©<U+FFFD>ˆë‹¤. <U+FFFD>°ì´<U+FFFD><U+FFFD> <U+FFFD>•ì‹<U+FFFD>€ Date <U+FFFD>´ì–´<U+FFFD><U+FFFD> <U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}
}
\value{
"\%Y-\%m-\%d" <U+FFFD>•ì‹<U+FFFD><U+FFFD> Date <U+FFFD>•ì‹ ë²¡í„°ê°€ ë°˜í™˜<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.
}
\description{
<U+FFFD>¤ëŠ˜<U+FFFD>¼ìë¥<U+FFFD> ê¸°ì<U+FFFD><U+FFFD>¼ë¡œ ìµœê·¼<U+FFFD>¼ìë¥<U+FFFD> <U+FFFD>‰ì¼ë§Œì„ ë½‘ì•„ <U+FFFD>´ê±°<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.
}
\examples{
recent_weekday(20)
recent_weekday(20, as.Date("2000-08-20"))
}
\seealso{
\link{is.weekend}
}

