% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{ror_diff}
\alias{ror_diff}
\title{Accounting k diff returns}
\usage{
ror_diff(xts, k, log.returns = F, start.omit = T, ...)
}
\arguments{
\item{xts}{time series dataset produced by \code{xts}.}

\item{k}{an integer vector.}

\item{log.returns}{logical asking whether to calculate log returns. Default is \code{FALSE}.}
}
\value{
\code{xts} about returns dataset.
}
\description{
Accounting k diff rate of returns
}
\examples{
ror_diff(sample_index, 1)
ror_diff(sample_index, 1, log.returns = T)

ror_diff(sample_index, 0:19)
ror_diff(sample_index, 0:19, log.returns = T)
}
