% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{svf2rds}
\alias{svf2rds}
\title{Write .rds about separated values file}
\usage{
svf2rds(file, ...)
}
\arguments{
\item{path}{a character vector; file paths}
}
\description{
구분자로 분리된 외부테이블 파일을 import 한 후 .rds 형태로 변환 및 압축하여 저장합니다. 저장되는 위치는 원본 외부테이블 파일이 있는 디렉토리입니다.
}
