% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{Object swap}
\usage{
swap(name1, name2, envir = parent.env(environment()))
}
\description{
Object swapping. (http://rosettacode.org/wiki/Generic_swap#R)
}
\examples{
x <- 1
y <- 2
swap("x", "y")
cat(x, y)
}
