% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmplot.R
\name{tmplot}
\alias{tmplot}
\title{ts and moving average plot for all fund items}
\usage{
tmplot(xts, mv = c(20, 60, 120), ...)
}
\arguments{
\item{xts}{an \code{xts} object}

\item{mv}{moving average parameters}

\item{...}{additional parameters passed to \code{facet_grid()}}

\item{plotly}{logical asking whether to using plotly class plotting}
}
\value{
\code{plotly} If TRUE, plotly class ploting. else ggplot2 class ploting. Default value is FALSE
}
\description{
ts and (three lines) moving average plot for all fund items
}
\examples{
library("quantmod")

data <- getSymbols("^ks11", from = Sys.Date() - 1000, to = Sys.Date(), auto.assign = F)

tmplot(data)
}
