% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trimming of whitespace in character}
\usage{
trim(cha, method = c("leading", "trailing"))
}
\arguments{
\item{cha}{ê°€ê³µì˜ <U+FFFD>€<U+FFFD>ì´ <U+FFFD>˜ëŠ” character type <U+FFFD>ë£Œë¥<U+FFFD> <U+FFFD>…ë ¥<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}

\item{method}{<U+FFFD>˜ë¼<U+FFFD>´ëŠ” ë°©ë²•<U+FFFD><U+FFFD> <U+FFFD>í•©<U+FFFD>ˆë‹¤.}
}
\description{
ê³µë<U+FFFD>(whitespace)<U+FFFD><U+FFFD> <U+FFFD>˜ë¼<U+FFFD>…ë‹ˆ<U+FFFD><U+FFFD>.
}
\examples{
trim("  Wow!  ", "leading")
trim("  Wow!  ", "trailing")
}

