% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which_approach}
\alias{which_approach}
\title{which approach}
\usage{
which_approach(x, k)
}
\arguments{
\item{x}{ì°¾ì„ <U+FFFD>€<U+FFFD>ì´ <U+FFFD>˜ëŠ” ë²¡í„°<U+FFFD>€<U+FFFD>…ì„ <U+FFFD>…ë ¥<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}

\item{k}{ì°¾ì•„<U+FFFD><U+FFFD> <U+FFFD><U+FFFD> ìµœê·¼<U+FFFD>‘ì <U+FFFD><U+FFFD> <U+FFFD>€<U+FFFD><U+FFFD> numeric <U+FFFD>•ë³´ë¥<U+FFFD> <U+FFFD>…ë ¥<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}
}
\description{
which approach index
}
\examples{
x <- runif(100)
k <- .5
which_approach(x, k)
}

