% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which_duplicated}
\alias{which_duplicated}
\title{Which duplicated all on data.frame}
\usage{
which_duplicated(x)
}
\description{
Which all duplicated index on data.frame
}
\examples{
index <- which_duplicated(iris)
iris[index, ]
}
