% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which_matrix}
\alias{which_matrix}
\title{which matrix}
\usage{
which_matrix(logical.matrix)
}
\arguments{
\item{logical.matrix}{logical ê°’ì´ <U+FFFD>´ê¸´ matrix type <U+FFFD><U+FFFD> <U+FFFD>…ë ¥ ë°›ìŠµ<U+FFFD>ˆë‹¤.}
}
\description{
which matrix index
}
\examples{
m <- matrix(sample(c(T,F), 30, T), 5, 6)
which(m)
which_matrix(m)
}

