% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_trim.R
\name{window}
\alias{window}
\title{Bandwidth indexing with Bound trim}
\usage{
window(W0, c, min, max)
}
\arguments{
\item{W0}{Bandwidth <U+FFFD>¬ê¸°ë¥<U+FFFD> <U+FFFD>¤ì •<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}

\item{c}{Bandwidth indexing <U+FFFD><U+FFFD> ê¸°ì<U+FFFD><U+FFFD>ì„ <U+FFFD>¤ì •<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.}

\item{min}{ìµœì†Œê²½ê³„ë¥<U+FFFD> ì§€<U+FFFD>•í•©<U+FFFD>ˆë‹¤.}

\item{max}{ìµœë<U+FFFD>ê²½ê³„ë¥<U+FFFD> ì§€<U+FFFD>•í•©<U+FFFD>ˆë‹¤.}
}
\description{
Bandwidth indexing with Bound trim
}
\examples{
window(10, 0, -20, 20)
window(10, 0, -5, 20)

}

