% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{xdiff_returns}
\alias{xdiff_returns}
\title{Accounting x diff returns}
\usage{
xdiff_returns(xts, x = 1, log.returns = F, start.omit = T)
}
\arguments{
\item{xts}{an \code{xts} object}

\item{x}{an integer value}

\item{log.returns}{logical asking whether to calculate log returns}
}
\value{
xts type about returns dataset
}
\description{
<U+FFFD>¼ê´„<U+FFFD>ìœ¼ë¡<U+FFFD> ëª¨ë“  x <U+FFFD>œì°¨ ê°’ì„ ê¸°ì<U+FFFD><U+FFFD>¼ë¡œ <U+FFFD>˜ìµë¥ ì„ ê³„ì‚°<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.
}
\examples{
xdiff_returns(sample_index, 1)
xdiff_returns(sample_index, 1, log.returns = T)
}

