% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{xdiffs_returns}
\alias{xdiffs_returns}
\title{Accounting x diff returns}
\usage{
xdiffs_returns(xts, x, log.returns = F, start.omit = T)
}
\arguments{
\item{xts}{an \code{xts} object}

\item{x}{an integer vector}

\item{log.returns}{logical asking whether to calculate log returns}
}
\value{
xts type about returns dataset
}
\description{
ê°ê¸° <U+FFFD>¤ë¥¸ x <U+FFFD>œì°¨ ê°’ì„ ê¸°ì<U+FFFD><U+FFFD>¼ë¡œ <U+FFFD>˜ìµë¥ ì„ ê³„ì‚°<U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>. <U+FFFD>°ë¼<U+FFFD><U+FFFD> \code{x}<U+FFFD><U+FFFD> \code{n = nrow(xts)} ë§Œí¼<U+FFFD><U+FFFD> ë²¡í„°<U+FFFD>€<U+FFFD>…ê°’<U+FFFD><U+FFFD> ë°›ì•„<U+FFFD><U+FFFD> <U+FFFD>©ë‹ˆ<U+FFFD><U+FFFD>.
}
\examples{
xdiffs_returns(sample_index, x = 0:19)
xdiffs_returns(sample_index, x = 0:19, log.returns = T)
}

