% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_trim.R
\name{bound_trim}
\alias{bound_trim}
\title{Bound trim the sequence vector}
\usage{
bound_trim(vector, min, max)
}
\arguments{
\item{vector}{input sequence. e.g \code{1:100}}
}
\description{
Bound trim the sequence vector
}
\examples{
bound_trim(1:100, 5, 95)
bound_trim(c(1, 5, 9, 10), 0, 9)
bound_trim(rnorm(100), -1.96, 1.96)
}
