% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Describe dataset using psych::describe() function}
\usage{
desc(x, choice = c("vars", "n", "mean", "sd", "median", "min", "max",
  "range"), ...)
}
\arguments{
\item{x}{A data frame or matrix}

\item{choice}{}

\item{...}{Arguments to be passed to \code{psych::describe()}}
}
\description{
psych 패키지의 \code{describe()} 를 이용하여 알맞게 수정한 alias 함수 입니다.
}
\examples{
desc(mtcars)
desc(mtcars, boxplot = T)
bound_trim(c(1, 5, 9, 10), 0, 9)
bound_trim(rnorm(100), -1.96, 1.96)
}

