% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librarys.R
\name{librarys}
\alias{librarys}
\title{Loading of many packages (using vector type) before auto update packages}
\usage{
librarys(packages, auto.ins = FALSE, message = TRUE)
}
\arguments{
\item{packages}{불러올 패키지를 vector 형으로 작성합니다. 문자열로 작성 합니다.}

\item{auto.ins}{만약 불러올 패키지 중에 설치가 되어있지 않은 패키지가 있다면 자동으로 설치할 것인지 여부를 정합니다.}

\item{message}{패키지 로드 결과 summary 를 출력할 것인지 여부를 정합니다.}
}
\value{
summary 결과를 data.frame 형으로 반환하며, \code{message=FALSE} 시 반환되는 것은 없습니다.
}
\description{
2개 이상의 Package들을 한번에 로드하며, 동시에 업데이트 가능 패키지를 업데이트 한 후 이전버전과 설치된 버전을 요약하여 출력합니다.
}
\examples{
# 단순 패키지 로드
librarys(packages=c("abind", "acepack", "animation", "arules", "ash"))
# 미설치 패키지 설치와 동시에 로드
librarys(packages=c("abind", "acepack", "animation", "arules", "ash"), auto.ins=TRUE)
}
\seealso{
\link{library}
}

