% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missmap.R
\name{missmap2}
\alias{missmap2}
\title{Missingness Map after sampling}
\usage{
missmap2(x, rep = T, ...)
}
\arguments{
\item{x}{A dataframe}
}
\description{
Plots a Missingness Map showing after sampling
}
\examples{
d <- matrix(rnorm(25), 5)
d[c(2, 4), c(1, 5)] <- NA
as.data.frame(d) \%>\% missmap2
}
