% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_paste0.R
\name{NA_paste0}
\alias{NA_paste0}
\title{Ignore NA paste function}
\usage{
NA_paste0(...)
}
\description{
결측치 NA 가 존재할 때, NA를 문자열로 인식하지 않고 공란으로 paste 를 수행합니다.
}
\examples{
### ex1
NA_paste0(c(3,6,NA,NA), c(NA,6,9,NA))

### ex2
A <- "Hello "
B <- c("world", "world!", NA)
paste0(A, B)
NA_paste0(A, B)
}
\seealso{
\link{paste0}
}

