% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste2.R
\name{paste2}
\alias{paste2}
\title{Ignore NA paste function}
\usage{
paste2(...)
}
\description{
Ignore NA paste function
}
\examples{
A <- "Hello "
B <- c("world", "world!", NA)
paste0(A, B)
paste2(A, B)
}
\seealso{
\link{paste0}
}
