% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_week.R
\name{recent_weekday}
\alias{recent_weekday}
\title{Sequence Date Generation (only Weekday) from recent day}
\usage{
recent_weekday(recent_day, to = Sys.Date())
}
\arguments{
\item{recent_day}{몇일차를 열거할 것인지 정합니다. 데이터 형식은 정수이어야 합니다.}

\item{to}{오늘일자기준을 정합니다. 데이터 형식은 Date 이어야 합니다.}
}
\value{
"\%Y-\%m-\%d" 형식인 Date 형식 벡터가 반환됩니다.
}
\description{
오늘일자를 기준으로 최근일자를 평일만을 뽑아 열거합니다.
}
\examples{
recent_weekday(20)
recent_weekday(20, as.Date("2000-08-20"))
}
\seealso{
\link{is.weekend}
}

