% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_week.R
\name{recent_weekend}
\alias{recent_weekend}
\title{Sequence Date Generation (only Weekend) from recent day}
\usage{
recent_weekend(recent_day, to = Sys.Date())
}
\arguments{
\item{recent_day}{input integer value}

\item{to}{recent day. default is \code{Sys.Date()}}
}
\value{
"\%Y-\%m-\%d" Date class
}
\description{
Sequence Date Generation (only Weekend) from recent day
}
\examples{
recent_weekend(8)
recent_weekend(8, as.Date("2000-08-20"))
}
\seealso{
\link{is.weekend}
}
