% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmatch.R
\name{regmatch}
\alias{regmatch}
\title{Match string using Regex patterns}
\usage{
regmatch(pattern, text, ...)
}
\description{
Match string out directly
}
\examples{
t <- c("Ab80", "bU09", "CD11", "DA01", "EW", "fP39")
regmatch("^[A-Z]*", t)
}
