% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmatch.R
\name{regmatch}
\alias{regmatch}
\title{Match string using Regex patterns}
\usage{
regmatch(pattern, text)
}
\description{
Match string using Regex patterns
}
\examples{
t <- c("A80", "B09", "C11", "D01", "E", "f39")
regmatch("[A-Z]", t)
}
