% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{rendering}
\alias{rendering}
\title{Open files after rmarkdown::render()}
\usage{
rendering(pattern = NULL, ...)
}
\description{
Open files after rmarkdown::render(..., encoding = "UTF-8") about \code{.rmd} format script.
}
\examples{
rendering()
}
