% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{retention}
\alias{retention}
\title{Retention}
\usage{
retention(x, by = c("days", "weeks", "months"), all = TRUE,
  percent_scale = TRUE)
}
\arguments{
\item{x}{input connect log data frame}

\item{all}{if FALSE, first date retention}
}
\description{
Retention
}
\examples{
set.seed(1004)
library(lubridate)
connect_log <- tibble(
  id = c(1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 3, 4, 4, 4, 5, 5),
  login_date = sample(seq(ymd_h(2018010100), ymd_h(2018010323), by = "hour"), 20)
) \%>\%
  arrange(id, login_date) # must be - first col : index, second col : date

retention(connect_log, by = "days")
retention(connect_log, by = "days", percent_scale = F)
}
