% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trimming of whitespace in character}
\usage{
trim(cha, method = "trailing")
}
\arguments{
\item{cha}{가공의 대상이 되는 character type 자료를 입력합니다.}

\item{method}{잘라내는 방법을 택합니다.}
}
\description{
공란(whitespace)을 잘라냅니다.
}
\examples{
trim("  Wow!  ", method="leading")
trim("  Wow!  ", method="trailing")
}

