% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trimming of whitespace in character}
\usage{
trim(cha, method = c("leading", "trailing"))
}
\arguments{
\item{cha}{input character type}

\item{method}{trimming method}
}
\description{
Trimming of whitespace in character
}
\examples{
trim("  Wow!  ", "leading")
trim("  Wow!  ", "trailing")
}

