% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which.matrix}
\alias{which.matrix}
\title{which matrix}
\usage{
which.matrix(logical.matrix)
}
\arguments{
\item{logical.matrix}{logical 값이 담긴 matrix type 을 입력 받습니다.}
}
\description{
which matrix index
}
\examples{
m <- matrix(sample(c(T,F), 30, T), 5, 6)
which(m)
which.matrix(m)
}

