% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which_approach}
\alias{which_approach}
\title{which approach}
\usage{
which_approach(x, k)
}
\arguments{
\item{x}{찾을 대상이 되는 벡터타입을 입력합니다.}

\item{k}{찾아야 할 최근접점에 대한 numeric 정보를 입력합니다.}
}
\description{
which approach index
}
\examples{
x <- runif(100)
k <- .5
which_approach(x, k)
}

