% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which_matrix}
\alias{which_matrix}
\title{which matrix}
\usage{
which_matrix(logical.matrix)
}
\description{
which matrix index
}
\examples{
m <- matrix(sample(c(T,F), 30, T), 5, 6)
which(m)
which_matrix(m)
}
