% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_trim.R
\name{window}
\alias{window}
\title{Bandwidth indexing with Bound trim}
\usage{
window(W0, c, min, max)
}
\arguments{
\item{W0}{Bandwidth size}

\item{c}{Bandwidth center point}
}
\description{
Bandwidth indexing with Bound trim
}
\examples{
window(10, 0, -20, 20)
window(10, 0, -5, 20)
}
